requirements.txt
django
djangorestframework
celery
flower
django-celery-beat
requests
redis
psycopg2-binary